testing main functionality
